Implemented functionality to sort Ingredient
Updated MainWindow to initialize the sort criteria combobox to store a TypeComparator and NameComparator. • Added sortIngredient() to be triggered when a sort criteria is selected to sort the list of ingredients using the currently selected sort criteria. Updated adding and removing ingredients with sortIngredientList to ensure the list is sorted after it has been updated.